Clippy improvements

+++ src/interaction.rs
@@ -63,7 +63,7 @@ pub fn get_cursor_position(
     // check if the cursor is inside the window and get its position
     if let Some(screen_pos) = wnd.cursor_position() {
         // get the size of the window
-        let window_size = Vec2::new(wnd.width() as f32, wnd.height() as f32);
+        let window_size = Vec2::new(wnd.width(), wnd.height());
         // convert screen position [0..resolution] to ndc [-1..1] (gpu coordinates)
         let ndc = (screen_pos / window_size) * 2.0 - Vec2::ONE;
         // matrix for undoing the projection and camera transform
